refactoring: Header 컴포넌트
* 컴포넌트 이름 변경 Header -> AppHeader
* 해당 컴포넌트 및 하위 컴포넌트를 src/App/... 로 이동
