+++ frontend/host-app/src/App/AppHeader/HeaderAccountAvatar.js
