+++ frontend/host-app/src/App/AppHeader/HeaderConfigAvatar.js
